pyproject.toml
[tool.poetry]
name = "authservice"
version = "0.1.0"
description = "service for auth"
authors = ["alenameshcheriakova"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
logger = "^1.4"
alembic = "^1.13.2"
pytest = "^8.3.2"
pydantic-settings = "^2.4.0"
grpcio = "^1.65.5"
grpcio-tools = "^1.65.5"
protobuf = "^5.27.3"
psycopg2-binary = "^2.9.9"
pytest-dotenv = "^0.5.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
